Do not start nvmf/iscsi target on behalf of initiator utility.
Move legacy mayastor start/stop functions to app module.
Move log related functions to logger module.
Set log level for rust unit tests to debug (instead of trace).
spdklog module is not used but left in the source tree just in case.
